Added support for clearing all cursors in one shot, restore cursors after loading project/data file.

+++ device.generic/src/main/java/nl/lxtreme/ols/device/generic/GenericDeviceController.java
@@ -90,7 +90,7 @@ public class GenericDeviceController implements DeviceController
 
           LOG.log( Level.FINE, "Progress {0}%", progress );
 
-          aCallback.updateProgress( progress );
+          aCallback.updateProgress( progress.intValue() );
         }
         else if ( GenericDevice.PROP_CAPTURE_STATE.equals( propertyName ) )
         {
